24 cleanup add an auto formatter and linter and other maintenance infrastructure

+++ 3DRenderer/display/tests/_viewport_2.cpp
@@ -1,39 +1,38 @@
-#include "CppUnitLite/TestHarness.h"
-#include "display/viewport.h"
-#include "display/frame_buffer.h"
+#include <chrono>
+#include <cmath>
 #include <iostream>
 #include <string>
-#include <cmath>
-#include <chrono>
 
+#include "CppUnitLite/TestHarness.h"
+#include "display/frame_buffer.h"
+#include "display/viewport.h"
 
-FrameBuffer create_circle_framebuffer(int width, int height, int cx, int cy, int radius) {
+FrameBuffer create_circle_framebuffer(int width, int height, int cx,
+                                      int cy, int radius) {
     FrameBuffer fb(width, height);
-    
+
     // Fill with white background
     for (size_t y = 0; y < height; y++) {
         for (size_t x = 0; x < width; x++) {
             fb.setPixel(x, y, 255, 255, 255);
         }
     }
-    
+
     // Draw green circle
     for (size_t y = 0; y < height; y++) {
         for (size_t x = 0; x < width; x++) {
-            int dx = x - cx;
-            int dy = y - cy;
-            if (dx*dx + dy*dy <= radius*radius) {
+            int dx = static_cast<int>(x) - cx;
+            int dy = static_cast<int>(y) - cy;
+            if (dx * dx + dy * dy <= radius * radius) {
                 fb.setPixel(x, y, 0, 255, 0);  // Green
             }
         }
     }
-    
+
     return fb;
 }
 
-
-TEST(Viewport, SetFrameDisplaysGreenCircle)
-{
+TEST(Viewport, SetFrameDisplaysGreenCircle) {
     int width = 600;
     int height = 400;
 
@@ -42,7 +41,8 @@ TEST(Viewport, SetFrameDisplaysGreenCircle)
     Viewport v1(width, height);
     v1.start();
 
-    v1.setFrame(create_circle_framebuffer(width, height, width / 2, height / 2, 100));
+    v1.setFrame(create_circle_framebuffer(width, height, width / 2,
+                                          height / 2, 100));
     v1.update();
 
     std::string answer;
@@ -54,9 +54,7 @@ TEST(Viewport, SetFrameDisplaysGreenCircle)
     v1.join();
 }
 
-
-TEST(Viewport, AnimatedCircleMovesDownward)
-{
+TEST(Viewport, AnimatedCircleMovesDownward) {
     int width = 600;
     int height = 400;
 
@@ -65,8 +63,9 @@ TEST(Viewport, AnimatedCircleMovesDownward)
     Viewport v1(width, height);
     v1.start();
 
-    for(int n = 0; n < 50; n++) {
-        v1.setFrame(create_circle_framebuffer(width, height, width / 2, height / 2 + n*4, 100));
+    for (int n = 0; n < 50; n++) {
+        v1.setFrame(create_circle_framebuffer(width, height, width / 2,
+                                              height / 2 + n * 4, 100));
         v1.update();
         std::this_thread::sleep_for(std::chrono::milliseconds(50));
     }
@@ -79,7 +78,6 @@ TEST(Viewport, AnimatedCircleMovesDownward)
 
     v1.join();
 }
-
 
 int main() {
     TestResult tr;
